rename metronome files and folders
