+++ src/components/MainWrapper/MainWrapper.tsx
@@ -1,5 +1,3 @@
-import TopNavBar from '../TopNavBar/TopNavBar';
-import PlaybackControls from '../PlaybackControls/PlaybackControls';
 import MenuWrapper from '../Menu/MenuWrapper.tsx';
 import './MainWrapper.css';
 import { useState } from 'react';
@@ -17,30 +15,19 @@ function MainWrapper({
   onReset,
   selectedStickings,
 }: Props) {
-  const [bpm, setBpm] = useState('90');
-  const [repeats, setRepeats] = useState('4');
+  // const [bpm, setBpm] = useState('90');
+  // const [repeats, setRepeats] = useState('4');
 
-  const handleBpmChange = (numberInput: string): void => {
-    setBpm(numberInput);
-  };
+  // const handleBpmChange = (numberInput: string): void => {
+  //   setBpm(numberInput);
+  // };
 
-  const handleRepeatChange = (numberInput: string): void => {
-    setRepeats(numberInput);
-  };
+  // const handleRepeatChange = (numberInput: string): void => {
+  //   setRepeats(numberInput);
+  // };
 
   return (
     <main>
-      <TopNavBar
-        bpmValue={bpm}
-        repeatValue={repeats}
-        onBpmChange={handleBpmChange}
-        onRepeatChange={handleRepeatChange}
-      />
-      <PlaybackControls
-        bpmValue={bpm}
-        repeatValue={repeats}
-        selectedStickings={selectedStickings}
-      />
       <MenuWrapper
         displayMenu={displayMenu}
         onFormChange={onFormChange}
