+++ src/components/MetronomeControls/MetronomeControls.css
