+++ src/components/MetronomeControls/MetronomeControls.tsx
